multitask_lightning/requirements.txt
albumentations==1.1.0
click==8.0.3
matplotlib==3.4.3
mean-average-precision @ git+https://github.com/bes-dev/mean_average_precision.git@9f4d65036de9c5deca7ec0c13c4d4278ecf8e3c3
numpy==1.21.2
opencv-python==4.5.3.56
pytorch-lightning==1.4.9
pytorch-toolbelt==0.4.4
PyYAML==5.4.1
segmentation-models-pytorch==0.2.0
tensorboard==2.6.0
torch==1.9.0
torchmetrics==0.5.1
torchvision==0.10.0
tqdm==4.62.3
coremltools==5.0b5
timm==0.4.12